rename lowecase and remove unused import
