+++ src/components/Blog/style/variant/blog.variant.tsx
