+++ src/components/Disclaimer/styles/info.style.module.css
