+++ src/components/Header/index.tsx
@@ -1,6 +1,5 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import {
-  faBarsStaggered,
   faCartShopping,
   faUser,
 } from '@fortawesome/free-solid-svg-icons';
