+++ src/components/Header/style/variant/header.variant.tsx
@@ -1,4 +1,4 @@
-import { table } from 'console';
+
 
 // nav list bottom
 export const listVariant = {
