+++ src/components/Help/style/help.styled.components.tsx
@@ -1,5 +1,3 @@
-import { faFontAwesome } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { motion } from 'framer-motion';
 import styled from 'styled-components';
