+++ src/components/Help/style/variant/help.variant.tsx
@@ -1,4 +1,3 @@
-import { animations } from 'framer-motion';
 
 export const hoverVariant = {
   hover: {
