+++ src/components/Slideshow/index.tsx
@@ -1,4 +1,3 @@
-import { time } from 'console';
 import {
   motion,
   useAnimation,
